Added API for PWM. Added button "Save" on the Select.shtlm page.

+++ Core/Inc/db.h
@@ -26,24 +26,24 @@ struct dbCron {
 };
 
 struct dbPinsConf {    // Создали структуру с необходимым набором типов элиментов.
-	uint8_t topin;		// Type of pins: 0 - NONE; 1 - BUTTON; 2 - RELAY; 3 - SWITCH;  4 - 1-WIRE; 5 - PWM; 6 - I2C;
+	int topin;		// Type of pins: 0 - NONE; 1 - BUTTON; 2 - RELAY; 3 - SWITCH;  4 - 1-WIRE; 5 - PWM; 6 - I2C;
 	int pwm;		// PWM frequency
-	uint8_t on;			// ON
-	uint8_t istate;		// Invert state
+	int on;			// ON
+	int istate;		// Invert state
 	int dvalue;		// Dimmer value
-	uint8_t ponr;		// Power on restore
+	int ponr;		// Power on restore
 	int ptype; 		// Pullup type 0 - NONE; 1 - GPIO_PULLUP; 2 - GPIO_PULLDOWN
-	uint8_t sclick; 	// SINGLE CLICK
+	int sclick; 	// SINGLE CLICK
 	char dclick[255]; 	// DOUBLE CLICK
 	char lpress[255]; 	// LONG PRESS
-	uint8_t numdevices; 	// Indicate the number of ds18b20 sensors
+	int numdevices; 	// Indicate the number of ds18b20 sensors
 //	int dcinter;	// Double-click interval
 //	int pclick; 	// Prevent Click
 	char info[30];	// Info
-	uint8_t onoff;		// On | Off
-	uint8_t event;     // Event (StateChanged - 0, Pressed -1, Released - 2, Repeat - 3, Hold - 4, LongClick - 5, Click - 6, DoubleClick - 7)
-	uint8_t act;       // Action (No action - 0, On - 1,  Off - 2, Toggle - 3, Value - 4, IncValue - 5)
-	uint8_t parametr;  // Parameter value for dimmer (0-255), value for IncValue (-255 to 255).
+	int onoff;		// On | Off
+	int event;     // Event (StateChanged - 0, Pressed -1, Released - 2, Repeat - 3, Hold - 4, LongClick - 5, Click - 6, DoubleClick - 7)
+	int act;       // Action (No action - 0, On - 1,  Off - 2, Toggle - 3, Value - 4, IncValue - 5)
+	int parametr;  // Parameter value for dimmer (0-255), value for IncValue (-255 to 255).
 	int timeout;     // Timeout (ms)
 	char condit[30]; // Condition
